pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-tax-toolkit"
version = "0.1.0"
description = "Automate tax compliance for AI agents earning revenue"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Venture", email = "agent@clawboard.io"}
]
dependencies = [
    "fastapi>=0.100.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
agent-tax = "agent_tax_toolkit.cli:main"
